Added execution base class and cleanup
