+++ src/citrine/informatics/design_candidate.py
@@ -159,9 +159,6 @@ class HierarchicalDesignMaterial(Serializable["HierarchicalDesignMaterial"]):
     mixtures = properties.Mapping(properties.UUID, properties.Object(Mixture), 'mixtures')
     """:Dict[UUID, Mixture]: mapping from Citrine ID to components the material is composed of"""
 
-    def __init__(self):
-        pass  # pragma: no cover
-
 
 class DesignCandidate(Serializable["DesignCandidate"]):
     """A candidate material generated by the Citrine Platform.
@@ -181,9 +178,6 @@ class DesignCandidate(Serializable["DesignCandidate"]):
     material = properties.Object(DesignMaterial, 'material')
     """:DesignMaterial: the material returned by the design workflow"""
 
-    def __init__(self):
-        pass  # pragma: no cover
-
 
 class HierarchicalDesignCandidate(Serializable["HierarchicalDesignCandidate"]):
     """A hierarchical candidate material generated by the Citrine Platform.
@@ -201,9 +195,6 @@ class HierarchicalDesignCandidate(Serializable["HierarchicalDesignCandidate"]):
     material = properties.Object(HierarchicalDesignMaterial, "material")
     """:HierarchicalDesignMaterial: the material returned by the design workflow"""
 
-    def __init__(self):
-        pass  # pragma: no cover
-
 
 class SampleSearchSpaceResultCandidate(Serializable["SampleSearchSpaceResultCandidate"]):
     """A hierarchical candidate material generated by the Citrine Platform.
@@ -217,6 +208,3 @@ class SampleSearchSpaceResultCandidate(Serializable["SampleSearchSpaceResultCand
     """:UUID: unique external Citrine ID of the execution"""
     material = properties.Object(HierarchicalDesignMaterial, "material")
     """:HierarchicalDesignMaterial: the material returned by the design workflow"""
-
-    def __init__(self):
-        pass  # pragma: no cover
